arnergy-ems_flasher-5285d8c804de/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
ntplib==0.4.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyserial==3.5
pytz==2021.1
requests==2.26.0
urllib3==1.26.7
serial.tools
